Clean up the comparison code
