+++ community/cypher/compatibility-suite/src/test/scala/cypher/GlueSteps.scala
@@ -187,61 +187,3 @@ object GlueSteps {
   val EXPECT_RESULT = "^the result should be:$"
 
 }
-
-
-case class represent(expected: util.List[util.Map[String, AnyRef]])
-  extends Matcher[util.List[util.Map[String, AnyRef]]] {
-
-  override def apply(actual: util.List[util.Map[String, AnyRef]]): MatchResult = {
-    MatchResult(matches = compare(expected, actual), "a mismatch found", "no mismatches found")
-  }
-
-
-  def compare(expected: util.List[util.Map[String, AnyRef]], actual: util.List[util.Map[String, AnyRef]]): Boolean = {
-    val expSorted = expected.asScala.toVector.map(_.asScala).sortBy(_.hashCode())
-    val actSorted = actual.asScala.toVector.map(_.asScala).sortBy(_.hashCode())
-
-    val bools = expSorted.zipWithIndex.map { case (expMap, index) =>
-      val b = cypherEqual(expSorted(index), actSorted(index))
-      if (!b)
-        println(s"not equal: $expMap and ${actSorted(index)}")
-      b
-    }
-    bools.reduce(_ && _)
-  }
-
-
-  def cypherEqual(expected: mutable.Map[String, AnyRef], actual: mutable.Map[String, AnyRef]) = {
-    val keys = expected.keySet == actual.keySet
-    val map = expected.keySet.map { key =>
-      cypherEqual2(expected(key), actual(key))
-    }
-    keys && map.reduce(_ && _)
-  }
-
-  def nodeEquals(expected: Node, actual: Node): Boolean = {
-    val labels = cypherEqual2(expected.getLabels, actual.getLabels)
-    val properties = cypherEqual2(expected.getAllProperties, actual.getAllProperties)
-
-    labels && properties
-  }
-
-  def relEquals(expected: Relationship, actual: Relationship): Boolean = ???
-
-  def pathEquals(expected: Path, actual: Path): Boolean = ???
-
-  def listEquals(expected: Seq[AnyRef], actual: Seq[AnyRef]): Boolean = ???
-
-  def mapEquals(expected: Map[String, AnyRef], actual: Map[String, AnyRef]): Boolean = ???
-
-  def cypherEqual2(expected: AnyRef, actual: AnyRef): Boolean = expected match {
-    case null => actual == null
-    case n: Node => nodeEquals(n, actual.asInstanceOf[Node])
-    case r: Relationship => relEquals(r, actual.asInstanceOf[Relationship])
-    case p: Path => pathEquals(p, actual.asInstanceOf[Path])
-    case l: Seq[_] => listEquals(l.asInstanceOf[Seq[AnyRef]], actual.asInstanceOf[Seq[AnyRef]])
-    case m: Map[_, _] => mapEquals(m.asInstanceOf[Map[String, AnyRef]], actual.asInstanceOf[Map[String, AnyRef]])
-    case _ => expected == actual
-  }
-
-}
